# Bug/performance fixes in sudoko.py: