Bug/performance fixes in sudoko.py:

1. Input parsing was botching handling of '0'
   (which presumably is supposed to mean the same thing as '.').
   To demonstrate (before this commit):
     python2 sudoku.py 1.....7....71.9...68..7......1.9.6.....3...2..4......3..8.6.1..5......4......2.05
   Notice the output contains a 0, which makes no sense.

2. Input parsing was doing islice on a list comprehension,
   which is pointless; changed it to do the islice on a
   generator expression instead, as intended.

3. The updating of prev,candidates was in the wrong place:
   it should be done after updating the board, not before it.
   Fixing this reduces unnecessary calls to eliminate_candidates,
   which improves the running time on my machine
   from 5.6s down to 3.8s for the default (Everest) board,
   and from 14s down to 10s for the "difficulty: off charts :)" example
   from collections/test_db.txt; that is:
     python2 sudoku.py 1.....7....71.9...68..7......1.9.6.....3...2..4......3..8.6.1..5......4......2..5

====================================================
Note: The above description is from the second commit d04bd94 of this pull request; the first commit 80fd632 is a previous pull request https://github.com/r1cc4rdo/sudoku/pull/2 which should be pulled first (if it's acceptable).  I'm not sure I'm managing these dependent pull requests properly.