@r1cc4rdo 2021-04-20
Agreed on all 3 points. Thanks.
I do remember having a reason for having eliminate_candidates at the end of the loop but that reason, whatever it was, is no more in the latest code I pushed and you modified.

Re: dependent PRs, you've pushed the same commit twice, in the previous PR and this.
There's no need to do that since the PR are not actually dependent on each other: in the first you ask me to pull a set of changes, in the second a superset of those. But what ends up in the PR is the diff with mainline in either case.