@donhatch 2021-04-20
Thanks for the pull!

Regarding the dependent PRs, while it's true the *content* of the two PRs was independent,
the *commit message* of the second does depend on the first (since it mentions example commands,
assume the first commit 80fd632 has happened).   In that sense it seemed that the second one has to come after the first, so that's why I didn't make the second one independent-- just in case you processed them in the opposite order, or you decided you didn't like the first one at all, I didn't want to risk having the commit message not make sense.   It's certainly awkward, in any case!